add gitignore

+++ src/Main.java
@@ -4,7 +4,7 @@ public class Main {
     public static void main(String[] args) {
 
         int  ticket = 3147;
-        double bonusMiles = 20;
+        int bonusMiles = 20;
         System.out.println("Количество начисленных миль:");
         System.out.println(ticket/bonusMiles); //Вывод бонусов
     }
